+++ Matjes/src/libraries_and_variables/basic_types.f90
@@ -19,7 +19,7 @@ type Op_real_order_N
 end type Op_real_order_N
 
 type Op_Im
-     complex(kind=16), pointer :: Op_loc(:,:)
+     complex(kind=8), pointer :: Op_loc(:,:)
 end type Op_Im
 
 !!!!!!!!!!!!!!!!!!!!!!!!!
